+++ app/Controllers/ExpenseCategoryController.php
@@ -59,6 +59,12 @@ class ExpenseCategoryController extends BaseController
         $Category = $model->where('id', $id)->first();
 
         if ($Category) {
+            if (!auth()->user()->can('expense-categories.edit'))
+                return $this->response->setJSON([
+                    'status' => false,
+                    'message' => "Don't have permission to edit this record!"
+                ]);
+
             if ($model->save($inputs)) {
                 $res = array_merge($res, [
                     'status' => true,
@@ -72,6 +78,12 @@ class ExpenseCategoryController extends BaseController
                 ]);
             }
         } else {
+            if (!auth()->user()->can('expense-categories.create'))
+                return $this->response->setJSON([
+                    'status' => false,
+                    'message' => "Don't have permission to create this record!"
+                ]);
+
             if ($model->save($inputs)) {
                 $res = array_merge($res, [
                     'status' => true,
@@ -105,6 +117,12 @@ class ExpenseCategoryController extends BaseController
      */
     public function delete($id = null)
     {
+        if (!auth()->user()->can('expense-categories.delete'))
+            return $this->response->setJSON([
+                'status' => false,
+                'message' => "Don't have permission to delete this record!"
+            ]);
+
         $model = new ExpenseCategoryModel();
         if ($model->delete($id)) {
             $res = [
